+++ e2e/utils/config.ts
@@ -1,0 +1,1 @@
+export const timeout = 30000
