--- e2e/utils/enter.ts
@@ -1,6 +1,0 @@
-export async function enter(input: WebdriverIO.Element) {
-  if (process.env.E2E_DEVICE === "ios") {
-    await input.sendKeys(["\n"])
-  }
-  await browser.pressKeyCode(66)
-}
