--- e2e/utils/go-back.ts
@@ -1,6 +1,0 @@
-export function goBack() {
-  if (process.env.E2E_DEVICE === "ios") {
-    return `//XCUIElementTypeButton[@name="Go back"]`
-  }
-  return `~Go back`
-}
