+++ e2e/utils/index.ts
@@ -1,6 +1,4 @@
-export * from "./selector"
-export * from "./swipe"
-export * from "./go-back"
-export * from "./enter"
-export * from "./scroll"
-export * from "./numberpad"
+export * from "./controls"
+export * from "./graphql"
+export * from "./use-cases"
+export * from "./config"
